Revert "change logout component to hook"
This reverts commit 077b225ace9e85f70bcc7090e42f2cccca0c0318.
