+++ src/component/user/AuthenticationButton.tsx
@@ -1,10 +1,8 @@
 import React, { Fragment, useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import useApi from '../../hook/useApi';
-import useLogout from '../../hook/useLogout';
 
 function AuthenticationButton() {
-  const { logoutHandler } = useLogout();
   const { sessionApi } = useApi();
   const [isLoggedIn, setIsLoggedIn] = useState(false);
 
@@ -15,9 +13,9 @@ function AuthenticationButton() {
       .catch(() => {
         setIsLoggedIn(false);
       });
-  }, [logoutHandler]);
+  }, []);
 
-  return <Fragment>{isLoggedIn ? <p onClick={logoutHandler}>Logout</p> : <Link to={'/login'}>Login</Link>}</Fragment>;
+  return <Fragment>{isLoggedIn ? <Link to={'/logout'}>Logout</Link> : <Link to={'/login'}>Login</Link>}</Fragment>;
 }
 
 export default AuthenticationButton;
